解决ReviewService.java的爆红问题，解决了合并冲突

+++ backend/src/main/java/com/foodadvisor/dto/IssueReviewVO.java
@@ -13,7 +13,7 @@ public class IssueReviewVO {
     private Long reviewId;
 
     /** 综合评分 1-5 */
-    private Integer rating;
+    private BigDecimal rating;
 
     /** 评价正文（前端按需截断） */
     private String content;
